requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.9
albumentations==1.4.8
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.7
dnspython==2.6.1
email_validator==2.1.1
environs==9.5.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.4
filelock==3.16.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
imageio==2.34.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.136
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
onnx==1.16.1
orjson==3.10.9
packaging==24.1
pandas==2.0.3
pillow==10.4.0
prettytable==3.10.0
propcache==0.2.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pymilvus==2.4.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.5
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.22.0
yarl==1.15.2
